re-do the rebase to improve subsequent merges
